V2

+++ cron/renew_certs.sh
@@ -1,5 +1,10 @@
 #!/bin/sh
 
+if [ "$DRY_RUN" = "true" ]; then
+  echo "Dry run is enabled"
+  exit 0
+fi
+
 cd /workdir
 echo "Renewing Let's Encrypt Certificates... (`date`)"
 docker compose run --rm --no-TTY --entrypoint certbot certbot renew --no-random-sleep-on-renew
